refactor(roslyn): split WorkspaceManager restore/analyzer subsystems (workspace-manager-decompose-restore-and-analyzer-subsystems)
Extracts the restore-staleness detection block and the unresolved-
analyzer-reference stripping logic out of the 1775-line WorkspaceManager
into two standalone, zero-external-caller collaborators:
RestoreStalenessDetector and UnresolvedAnalyzerReferenceStripper. Pure
code motion, mirroring the existing WorkspaceSessionLoader seam pattern
(internal class, public virtual methods, logger passed per call).
