+++ changelog.d/workspace-manager-decompose-restore-and-analyzer-subsystems.md
@@ -1,0 +1,5 @@
+---
+category: Maintenance
+---
+
+- **Maintenance:** Split WorkspaceManager's restore-staleness detection and unresolved-analyzer-reference isolation into two standalone collaborators (RestoreStalenessDetector, UnresolvedAnalyzerReferenceStripper). No behavior change.
